+++ src/__tests__/LoginContainer.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 
-import LoginContainer from '../components/pages/Login/LoginContainer';
+import LoginContainer from '../components/Login/LoginContainer';
 
 describe('<LoginContainer /> test suite', () => {
   test('signin widget mounts successfully', () => {
