+++ src/selectors/index.js
@@ -1,0 +1,1 @@
+const program = state => state.programReducer.programs_list;
